refactor!: extract the methods `Spec::each_element` and `Spec::any_element` into the trait `AssertElements`
