+++ src/iterator/tests.rs
@@ -1,4 +1,5 @@
 use crate::prelude::*;
+use crate::std::string::String;
 use crate::std::{vec, vec::Vec};
 
 #[derive(Debug)]
@@ -229,6 +230,199 @@ fn verify_custom_iterator_does_not_contain_fails() {
   expected: not 19
 "]
     );
+}
+
+mod all_elements {
+    use super::*;
+
+    #[derive(Debug)]
+    struct TestPerson {
+        name: String,
+        age: u8,
+    }
+
+    #[test]
+    fn assert_each_element_of_an_iterator_of_integer() {
+        let subject = [2, 4, 6, 8, 10];
+
+        assert_that(subject)
+            .is_not_empty()
+            .each_element(|e| e.is_positive().is_at_most(20));
+    }
+
+    #[test]
+    fn assert_each_element_of_an_iterator_of_person() {
+        let subject = vec![
+            TestPerson {
+                name: "John".into(),
+                age: 42,
+            },
+            TestPerson {
+                name: "Jane".into(),
+                age: 20,
+            },
+        ];
+
+        assert_that(subject)
+            .is_not_empty()
+            .each_element(|person| person.extracting(|p| p.age).is_at_most(42));
+    }
+
+    #[test]
+    fn assert_each_element_of_a_borrowed_iterator_of_integer() {
+        let subject = vec![2, 4, 6, 8, 10];
+
+        assert_that(&subject)
+            .is_not_empty()
+            .each_element(|e| e.is_positive().is_at_most(&20));
+    }
+
+    #[test]
+    fn assert_each_element_of_a_borrowed_iterator_of_person() {
+        let subject = vec![
+            TestPerson {
+                name: "John".into(),
+                age: 42,
+            },
+            TestPerson {
+                name: "Jane".into(),
+                age: 20,
+            },
+        ];
+
+        assert_that(&subject)
+            .is_not_empty()
+            .each_element(|person| person.extracting(|p| &p.name).starts_with('J'));
+    }
+
+    #[test]
+    #[should_panic = "expected numbers [1] to be not equal to 4\n   but was: 4\n  expected: not 4\n"]
+    fn assert_each_element_of_an_iterator_panics_if_one_assertion_fails() {
+        let subject = [2, 4, 6, 8, 10];
+
+        assert_that(subject)
+            .named("numbers")
+            .is_not_empty()
+            .each_element(|e| e.is_not_equal_to(4));
+    }
+
+    #[test]
+    fn verify_assert_each_element_of_an_iterator_fails() {
+        let subject = [2, 4, 6, 8, 10];
+
+        let failures = verify_that(&subject)
+            .named("numbers")
+            .each_element(|e| e.is_greater_than(&2).is_at_most(&7))
+            .display_failures();
+
+        assert_eq!(
+            failures,
+            &[
+                r"expected numbers [0] to be greater than 2
+   but was: 2
+  expected: > 2
+",
+                r"expected numbers [3] to be at most 7
+   but was: 8
+  expected: <= 7
+",
+                r"expected numbers [4] to be at most 7
+   but was: 10
+  expected: <= 7
+",
+            ]
+        );
+    }
+
+    #[test]
+    fn assert_any_element_of_an_iterator_of_str() {
+        let subject = ["one", "two", "three", "four", "five"];
+
+        assert_that(subject)
+            .is_not_empty()
+            .any_element(|e| e.contains("ee"));
+    }
+
+    #[test]
+    fn assert_any_element_of_an_iterator_of_person() {
+        let subject = vec![
+            TestPerson {
+                name: "John".into(),
+                age: 42,
+            },
+            TestPerson {
+                name: "Jane".into(),
+                age: 20,
+            },
+        ];
+
+        assert_that(subject)
+            .is_not_empty()
+            .any_element(|person| person.extracting(|p| p.age).is_at_most(20));
+    }
+
+    #[test]
+    fn assert_any_element_of_a_borrowed_iterator_of_str() {
+        let subject = vec!["one", "two", "three", "four", "five"];
+
+        assert_that(&subject)
+            .is_not_empty()
+            .any_element(|e| e.starts_with("fi"));
+    }
+
+    #[test]
+    fn assert_any_element_of_a_borrowed_iterator_of_person() {
+        let subject = vec![
+            TestPerson {
+                name: "John".into(),
+                age: 42,
+            },
+            TestPerson {
+                name: "Jane".into(),
+                age: 20,
+            },
+        ];
+
+        assert_that(&subject)
+            .is_not_empty()
+            .any_element(|person| person.extracting(|p| &p.name).ends_with('n'));
+    }
+
+    #[test]
+    fn verify_assert_any_element_of_an_iterator_fails() {
+        let subject = ["one", "two", "three", "four", "five"];
+
+        let failures = verify_that(subject)
+            .named("words")
+            .any_element(|e| e.starts_with("fu"))
+            .display_failures();
+
+        assert_eq!(
+            failures,
+            &[
+                r#"expected words [0] to start with "fu"
+   but was: "one"
+  expected: "fu"
+"#,
+                r#"expected words [1] to start with "fu"
+   but was: "two"
+  expected: "fu"
+"#,
+                r#"expected words [2] to start with "fu"
+   but was: "three"
+  expected: "fu"
+"#,
+                r#"expected words [3] to start with "fu"
+   but was: "four"
+  expected: "fu"
+"#,
+                r#"expected words [4] to start with "fu"
+   but was: "five"
+  expected: "fu"
+"#,
+            ]
+        );
+    }
 }
 
 mod filtered_elements {
